Refactored components into folders
