+++ components/footer/Footer.js
@@ -1,5 +1,5 @@
-import { H2, H3, P } from "../style/typography";
-import { getString } from "../lib/richText";
+import { H2, H3, P } from "../../style/typography";
+import { getString } from "../../lib/richText";
 export default function Footer({ footerData }) {
     const {
         contact_email: {url: contactEmail},
