+++ pages/archive/[page].js
@@ -1,7 +1,7 @@
 import React from "react";
 import { getStoriesByPage, getStoryArchivePage, getNavigation, getFooter } from "../../lib/api";
-import Navigation from "../../components/Navigation";
-import Footer from "../../components/Footer";
+import Navigation from "../../components/navigation/Navigation";
+import Footer from "../../components/footer/Footer";
 import { H1, H2, P } from "../../style/typography";
 import { getString } from "../../lib/richText";
