+++ pages/contact.js
@@ -1,6 +1,6 @@
 import { getContactPage, getNavigation, getFooter } from "../lib/api";
-import Navigation from "../components/Navigation";
-import Footer from "../components/Footer";
+import Navigation from "../components/navigation/Navigation";
+import Footer from "../components/footer/Footer";
 import { H1, P } from "../style/typography";
 import { getString } from "../lib/richText";
